move all logic with addin, removing todos and updating locale storage into the saga file, created page directory for TodoApp Page, created root styles.js file for all styles
