+++ src/App.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 
 // local dependencies
 import store from './redux/store';
-import ToDoApp from './components/ToDoApp';
+import ToDoApp from './pages/TodoApp/TodoApp';
 
 const App = () => (
   <Provider store={store}>
