--- src/components/PostList/PostList.scss
@@ -1,8 +1,0 @@
-.post-list {
-    max-width: 1200px;
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-    margin: 0 auto;
-    padding: 50px 20px;
-}
