+++ src/redux/reducer.js
@@ -1,4 +1,4 @@
-import { ADD_TODO, META, REMOVE_TODO } from './actions';
+import { META } from './actions';
 
 const initial_state = {
     todos: [],
@@ -7,22 +7,10 @@ const initial_state = {
 
 const reducer = (state=initial_state, action) => {
     switch (action.type) {
-        case ADD_TODO:
-            return {
-                ...state,
-                todos: [action.todo, ...state.todos],
-            }
-        
             case META:
                 return {
                     ...state,
                     ...action.payload,
-                }
-
-            case REMOVE_TODO:
-                return {
-                    ...state,
-                    todos: state.todos.filter(todo => todo.id !== action.id)
                 }
 
         default:
