+++ src/styles.js
@@ -1,0 +1,5 @@
+import 'bootstrap/dist/css/bootstrap.min.css'
+import './index.css';
+import './pages/TodoApp/TodoApp.scss';
+import './components/TodoList/TodoList.scss';
+import './components/TodoItem/TodoItem.scss';
